+++ source/renderer/app/api/api.js
@@ -1728,7 +1728,7 @@ export default class AdaApi {
   setTestingNewsFeed: (testingNewsFeedData: GetNewsResponse) => void;
   setTestingStakePools: (testingStakePoolsData: Array<StakePoolProps>) => void;
   setTestingWallets: (testingWalletsData: Array<WalletProps>) => void;
-  setTestingWallet: (testingWalletData: Object, walletIndex?: 0) => void;
+  setTestingWallet: (testingWalletData: Object, walletIndex?: number) => void;
 
   // Stake pools testing utility
   setFakeStakePoolsJsonForTesting: (
@@ -1747,9 +1747,9 @@ const _createWalletFromServerData = action(
       address_pool_gap: addressPoolGap,
       balance,
       name,
-      state,
       passphrase,
       delegation,
+      state: syncState,
       isLegacy = false,
     } = wallet;
 
@@ -1794,7 +1794,7 @@ const _createWalletFromServerData = action(
       passwordUpdateDate:
         passphraseLastUpdatedAt && new Date(passphraseLastUpdatedAt),
       hasPassword: passphraseLastUpdatedAt !== null,
-      syncState: state,
+      syncState,
       isLegacy,
       delegatedStakePoolId,
       delegationStakePoolStatus,