+++ source/renderer/app/themes/daedalus/light-blue.js
@@ -882,6 +882,20 @@ export const LIGHT_BLUE_THEME_OUTPUT = {
       'rgba(94, 96, 102, 0.15)',
     '--theme-wallet-settings-resync-button-label-color': 'rgba(94, 96, 102, 1)',
   },
+  walletNotRespondingOverlay: {
+    '--theme-wallet-not-responding-background-color': 'rgba(36, 62, 98, 0.96)',
+    '--theme-wallet-not-responding-button-background-color': '#2a4263',
+    '--theme-wallet-not-responding-button-background-color-hover': '#fafbfc',
+    '--theme-wallet-not-responding-button-border-color': '#fafbfc',
+    '--theme-wallet-not-responding-button-text-color': '#fafbfc',
+    '--theme-wallet-not-responding-button-text-color-hover': '#243e62',
+    '--theme-wallet-not-responding-description-background-color':
+      'rgba(0, 0, 0, 0.1)',
+    '--theme-wallet-not-responding-description-text-color': '#fafbfc',
+    '--theme-wallet-not-responding-icon-color': '#fafbfc',
+    '--theme-wallet-not-responding-link-text-color': '#fafbfc',
+    '--theme-wallet-not-responding-title-text-color': '#fafbfc',
+  },
 };
 
 const LIGHT_BLUE_THEME_PARAMS: CreateThemeParams = {